Step - 6, fixes tests
