+++ server/src/jest-globals.js
@@ -1,2 +1,3 @@
-jest.mock('./models');
+jest.mock('./models/index');
 jest.mock('./lib/notifications');
+jest.mock('./models-abstract/index', () => require('./models-abstract/__mocks__/models-abstract'));
